VDB-5400: skip the test if cannot find the binary

+++ test/internal/align-cache/runtest.sh
@@ -6,6 +6,11 @@ tool_binary=$3
 
 echo "testing ${tool_binary}"
 
+if ! test -f ${DIRTOTEST}/${tool_binary}; then
+    echo "${DIRTOTEST}/${tool_binary} does not exist. Skipping the test."
+    exit 0
+fi
+
 rm -rf CSRA_file.cache
 echo "vdb/schema/paths = \"${VDB_INCDIR}\"" > tmp.kfg
 output=$(VDB_CONFIG=`pwd` ${DIRTOTEST}/${tool_binary} \
