+++ tools/loaders/general-loader/test/vschema/runtestcase.sh
@@ -58,6 +58,11 @@ TEMPDIR=$WORKDIR/$SUITEID/actual/$CASEID
 
 echo "running test $SUITEID:$CASEID"
 
+if ! test -f ${LOAD}; then
+    echo "${LOAD} does not exist. Skipping the test."
+    exit 0
+fi
+
 mkdir -p $TEMPDIR
 rm -rf $TEMPDIR/*
 if [ "$?" != "0" ] ; then
